+++ Samples/2.3/UnitTesting/src/Grains.Interfaces/ICallingGrain.cs
@@ -1,8 +1,11 @@
+using System.Threading.Tasks;
 using Orleans;
 
 namespace Grains
 {
     public interface ICallingGrain : IGrainWithGuidKey
     {
+        Task IncrementAsync();
+        Task PublishAsync();
     }
 }
