--- Samples/2.3/UnitTesting/src/Grains.Interfaces/ICounterGrain.cs
@@ -1,16 +1,0 @@
-using System.Threading.Tasks;
-using Orleans;
-
-namespace Grains
-{
-    public interface ICounterGrain : IGrainWithStringKey
-    {
-        Task IncrementAsync();
-
-        Task<int> GetValueAsync();
-
-        Task SaveAsync();
-
-        Task PublishAsync();
-    }
-}
